Add solutions for 2022:7 "No Space Left On Device"
tried to solve this one using only incrementing sums, which
worked fine for the test input but not the actual puzzle input.

By a complete rewrite to actually render the tree as a map,
it worked. The spontanious data store is a mess and cost me
much time to work around, since I wanted to have a list of tuples
with all the sizes for each directory.

Took 2 days to figure this one out. Not proud.